:recycle: (pc-14462) create api folder and move api.ts
