+++ src/features/auth/signup/PhoneValidation/SetPhoneNumberDeprecated.tsx
@@ -10,7 +10,7 @@ import { QuitSignupModal } from 'features/auth/components/QuitSignupModal'
 import { Paragraphe } from 'features/auth/components/signupComponents'
 import { useAppSettings } from 'features/auth/settings'
 import { SignupStep } from 'features/auth/signup/enums'
-import { useSendPhoneValidationMutation } from 'features/identityCheck/api'
+import { useSendPhoneValidationMutation } from 'features/identityCheck/api/api'
 import { UseNavigationType } from 'features/navigation/RootNavigator'
 import { accessibilityAndTestId } from 'libs/accessibilityAndTestId'
 import { amplitude } from 'libs/amplitude'
